Docs : 커뮤니티 관련 컴포넌트 구조 통일
